requirements.txt
MetaTrader5>=5.0.0
pandas>=1.3.0
numpy>=1.20.0
pandas-ta
python-dotenv>=0.19.0
flask>=2.0.0
flask-socketio>=5.0.0
plotly>=5.0.0